requirements.txt
joblib==1.1.1
matplotlib==3.6.2
numpy==1.23.4
pandas==1.5.1
Pillow==9.3.0
scikit_learn==1.1.3
streamlit==1.14.1
